+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+
+(unreleased)
+------------
+* Add InMemory backend
+
+
 0.0.1 (2022-07-30)
 ------------------
 * First release on PyPI.